+++ RELEASES.md
@@ -1,5 +1,6 @@
 Version 0.9.5 (202X-XX-XX)
 ========================
+* Hyundai Azera 2022 support thanks to sunnyhaibin!
 * Hyundai Ioniq 6 2023 support thanks to sunnyhaibin, alamo3, and sshane!
 * Hyundai Kona Electric 2023 (Korean version) support thanks to sunnyhaibin and haram-KONA!
 * Kia Sorento Hybrid 2023 support thanks to sunnyhaibin!